REF: Moved modifier download code to downloader.py and renamed download.py -> utils.py
